refactor code according to the reviewer's comments
