+++ src/main/java/org/telran/codecrustpizza/dto/user/UserCreateRequestDto.java
@@ -1,8 +1,9 @@
 package org.telran.codecrustpizza.dto.user;
 
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.NotNull;
 
 public record UserCreateRequestDto(
 
@@ -13,18 +14,8 @@ public record UserCreateRequestDto(
         @Email
         String email,
 
-        @NotBlank
-        @Pattern(
-                regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,20}$",
-                message = "Password must be between 8 and 20 uppercase and lowercase symbols, digit and special symbols '@#$%^&+='"
-        )
-        String password,
-
-        @NotBlank
-        @Pattern(
-                regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\\S+$).{8,20}$",
-                message = "Password must be between 8 and 20 uppercase and lowercase symbols, digit and special symbols '@#$%^&+='"
-        )
-        String confirmPassword
+        @NotNull
+        @Valid
+        UserChangePasswordRequestDto passwordRequestDto
 ) {
 }
