+++ src/main/java/org/telran/codecrustpizza/entity/User.java
@@ -72,7 +72,8 @@ public class User {
     private Set<Address> addresses = new HashSet<>();
 
     @Enumerated(value = EnumType.STRING)
-    private Role role;
+    @Builder.Default
+    private Role role = Role.USER;
 
     private boolean isBlocked;
